Add Version Vector with Exceptions
This commit adds support for a form of causality barrier called a
version vector with exceptions (VVwE). It allows reception of causally
ordered messages in any order with low memory overhead. It is intended
to be used with CRDTs like LSEQ which require an ordering between
deletions and insertions. Specifically, a deletion must happen after the
corresponding insertion has already been seen.